webhook: remove deep clone of JobDesc in the worker.

+++ src/lib/handler.rs
@@ -113,19 +113,20 @@ async fn parse_and_handle(
 
                 let job = job::JobDesc::new(&action, &reviewer, &sha, pr_number, &head_ref);
                 let job_id  = job.id.clone();
+                let job = Arc::new(RwLock::new(job));
 
                 {
                     let mut job_registry = job_registry.write().await;
-                    job_registry.jobs.insert(job_id.clone(), Arc::new(RwLock::new(job)));
-                    let c_job: Arc<RwLock<job::JobDesc>>;
+                    job_registry.jobs.insert(job_id.clone(), job.clone());
                     let c_job_id: Uuid;
                     if let Some(c_id) = job_registry.running_jobs.get(head_ref.clone()) {
                         c_job_id = c_id.clone();
-                        if let Some(j) = job_registry.jobs.get(&c_job_id) {
-                            c_job = j.clone();
-                            let curr_job = c_job.read().await;
-                            if head_ref == curr_job.head_ref {
-                                info!("New job on same head_ref, killing current job");
+                        if let Some(curr_job) = job_registry.jobs.get(&c_job_id) {
+                            let curr_job = curr_job.read().await;
+                            // TODO(azhng): stop gap solution to avoid accidentally killing server
+                            //  process
+                            if head_ref == curr_job.head_ref && curr_job.pid != 0 {
+                                info!("New job on same head_ref, killing current job pid: {}", &curr_job.pid);
                                 // [To-Do] Fix, could potentially kill a process which is already dead
                                 signal::kill(Pid::from_raw(curr_job.pid as i32), Signal::SIGINT);
                             }
